+++ src/store/reducers/recipeReducer.tsx
@@ -33,19 +33,13 @@ const recipeSlice = createSlice({
   },
   extraReducers: (builder) => {
     builder.addCase(getRecipes.pending, (state: RecipeState) => {
-      console.log('Pending...');
       state.isFecthingRecipes = true;
     });
     builder.addCase(getRecipes.fulfilled, (state: RecipeState, action: any) => {
-      console.log('fulfilled');
       state.recipes = action.payload;
       state.isFecthingRecipes = false;
     });
     builder.addCase(getRecipes.rejected, (state: RecipeState, action: any) => {
-      console.log(
-        'Rejected What even is action.payload on rejected? --- ',
-        action.payload
-      );
       state.recipes = [];
       state.isFecthingRecipes = false;
     });
